--- Assets/CombiContentsUI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 55962b0aec08ee94297fd9f3a734b7f3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
