--- Assets/Game/Resources/Item.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7bc32e5f998c17e43b37e38d2d815ea6
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
