--- Assets/Game/Resources/Item/MaterialItem.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 029460e173da7854d884b20ecf1581ed
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
