--- Assets/Game/Resources/Item/MaterialItem/DearHorn.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9bc47dcd3268d194a8828e36d1c49704
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
