--- Assets/Game/Resources/Item/MaterialItem/DearLeather.asset
@@ -1,18 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: e0f38e00f7fc8d84caace9fb4999307e, type: 3}
-  m_Name: DearLeather
-  m_EditorClassIdentifier: 
-  itemName: "\uC0AC\uC2B4\uAC00\uC8FD"
-  contents: "\uC0AC\uC2B4\uC758 \uAC00\uC8FD"
-  value: 100
-  sprite: {fileID: 21300000, guid: 0045c0874948aad448456076260af71e, type: 3}
