--- Assets/Game/Resources/Item/MaterialItem/DearLeather.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3d36ab23b6ed91948a9b453e393b8508
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
