+++ Assets/Game/Resources/Json/EquipmentItemRecipeJson.json
@@ -1,0 +1,8 @@
+{
+  "materialA": [ "사슴뿔", "사슴뿔" ],
+  "materialB": [ "사슴가죽", "사슴가죽" ],
+  "numA": [ 3, 3 ],
+  "numB": [ 3, 3 ],
+  "gold": [ 1000, 500 ],
+  "result": [ "철검", "강철검" ]
+}
