+++ Assets/Game/Resources/Json/EquipmentItemRecipeJson.json.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8d17b76c6887c394aa152b4cf3d1f12e
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
