+++ Assets/Game/Scripts/Item/CombiContentsUI.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+using TMPro;
+
+public class CombiContentsUI : MonoBehaviour
+{
+    public Image image;
+    public TextMeshProUGUI textMeshProUGUI;
+}
