+++ Assets/Game/Scripts/Item/EquipmentCombinationSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d06c2697b73244449b6b124d2963b2e2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
