--- Assets/Game/Scripts/Item/ItemInformation.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0f13820210bd0be428847e5d358a3975
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
