--- Assets/Game/Scripts/Item/ScriptableMaterialItem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e0f38e00f7fc8d84caace9fb4999307e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
