+++ Assets/Game/Scripts/Item/UseItemCombinationSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 261650ea205222c4fb1f7cd35ed2d6fd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
