+++ Assets/Game/Scripts/Npc.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a6ead103478802f44a63bc58cf1ec3ac
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
